move stateVersionManager out of helpers
