+++ src/stream/stateVersionManager.ts
